Fix tox-py27 stable jobs of networking-bgpvpn
openstack-tox-py27 stable jobs fail for networking-bgpvpn for a couple
of weeks now. The problem according to the logs [1] [2]:

ERROR! /home/zuul/src/git.openstack.org/openstack/networking-odl not found
In Zuul v3 all repositories used need to be declared
in the 'required-projects' parameter on the job.
To fix this issue, add:
  openstack/networking-odl
to 'required-projects'.

On master tox_install.sh is already removed, but on stable branches it is
still there, and networking-odl should be installed with it, but somehow
it is failing in zuul.

[1] http://logs.openstack.org/periodic-stable/git.openstack.org/openstack/networking-bgpvpn/stable/queens/openstack-tox-py27/3e009ea/job-output.txt.gz#_2018-05-17_06_17_40_436575
[2] http://logs.openstack.org/periodic-stable/git.openstack.org/openstack/networking-bgpvpn/stable/pike/openstack-tox-py27/2d63b40/job-output.txt.gz#_2018-05-17_06_14_11_899842

Change-Id: I89fd8fc85bad9a7867e392b34d297644f76873ee