Remove direct tagging/branching for PowerVMStackers
The release process for official OpenStack components now uses
release requests using changes proposed to the openstack/releases
git repository. This allows peer review of release requests,
including version numbers, and run a series of checks to make
sure the release is likely to be successful once the tag is
pushed. The same process supports the creation of additional repo
branches as well. This all ensures that releases are reported
correctly on the releases.openstack.org website.

As a result of this process, nobody needs to push tags directly
or create branches directly. As an official OpenStack project
team, PowerVMStackers has switched to using that process already
for recent releases. However the ACLs were not fixed to remove
direct tagging/branching yet.

Change-Id: I7b9db2c47f226032527405b8d5b0fe29c825e205