Remove irrelevant-files from legacy-rally-dsvm-verify-full-discover-resources
This job has both files and irrelevant-files in its gate variant.
The check job only has files.  Having both is very confusing, and
most of the time just works out to be the files matcher anyway.
An upcoming change to zuul will prohibit having both on the same
variant, so remove the irrelevant-files from this one, as that's
likely what was intended anyway.

Change-Id: I49291b0702e468e4d947b9fc839bdf2a86d777a1