Remove direct tagging/branching for yaql
The release process for official OpenStack components now uses
release requests using changes proposed to the openstack/releases
git repository. This allows peer review of release requests,
including version numbers, and run a series of checks to make
sure the release is likely to be successful once the tag is
pushed. The same process supports the creation of additional repo
branches as well. This all ensures that releases are reported
correctly on the releases.openstack.org website.

As a result of this process, nobody needs to push tags directly
or create branches directly. The yaql library is associated with
the Murano team, which follows that process already. Let's fix the
ACLs to remove direct tagging/branching for yaql as well.

Change-Id: I98f438ebfa521bd59e3f607e99d73e2ab0777b91