Use tox for grafyaml validation
Currently this calls out to grafyaml-validate.sh which tries to
install zuul-cloner and then pull grafyaml from that.

Since zuulv3 doesn't *have* zuul-cloner, this means the job is broken
for local testing.  It works in the gate only because zuul-cloner
compatability shims are installed.

To maintain the testing-against-master behaviour, the dependent change
provides the grafyaml source in the environment; otherwise we go to
git directly.

Depends-On: https://review.openstack.org/572251
Change-Id: I3d5cd062414b77aaa70d2669730fda870ca0168e