Add afsmon project
This is an import of some code I developed for proof-of concept to
send AFS statistic to statsd.  It is currently working in a manual
fashion (see some details on the stats in
I8066f05df0c8eb2f4e211beaa25bccbd0261ca9b) so I think this fits best
in openstack-infra before it gets puppeted for automatic deployment.

Needed-By: https://review.openstack.org/572527
Change-Id: Ie7c491acf4ff734964be73487dc230478adfa96f