Add kata-run job to kata-containers/proxy
This is likely to not work yet, but since we aren't gating on
kata-containers/proxy we won't import job config from that repo which
would make these changes self testing. In order to get the ball moving
we need to start somewhere so start here.

This adds a vexxhost specific bionic image which we will start with
until the nodepool azure driver is tested and deployed. We should be
able to make progress on adding jobs and getting the driver going in
parallel this way.

Change-Id: Ia56c2cb3046e31e50014c8bf24342cceacb7db8a