Remove direct tagging/branching for QA repos
The release process for official OpenStack components
now uses release requests using changes proposed to the
openstack/releases git repository. This allows peer review
of release requests, including version numbers, and run a
series of checks to make sure the release is likely to be
successful once the tag is pushed. The same process
supports the creation of additional repo branches as well.
This all ensures that releases are reported correctly on
the releases.openstack.org website.

As a result of this process, nobody needs to push tags
directly or create branches directly. There were a number
of old QA deliverables with stale ACLs, let's fix them.

Change-Id: I5a098095d87945517fb066a9aa125a481a96118a