Fix ACLs for tripleo-ci and dib-utils
The release process for official OpenStack components now uses
release requests using changes proposed to the openstack/releases
git repository. This allows peer review of release requests,
including version numbers, and run a series of checks to make
sure the release is likely to be successful once the tag is
pushed. The same process supports the creation of additional repo
branches as well. This all ensures that releases are reported
correctly on the releases.openstack.org website.

As a result of this process, nobody needs to push tags directly
or create branches directly. tripleo-ci and dib-utils are under
the TripleO team and should be fixed.

For dib-utils, the simplest is to switch it to using the tripleo
ACLs, since it's now owned by that team.

Change-Id: I5f40f2318fe3a15fca630e32c2dfcb73104ec403