Build wheels against constraints
I was looking at wheel build failures, and in some cases builds fail
because they're not constrained to versions already in the
constraints.

For example; pylxd failed to build a wheel [1] due to [2], but
upper-constraints.txt would have kept urllib to 1.22.  Similarly
several django plugins fail on python2 as they try to bring in Django2
which only works with python3.

Pass the upper-constraints.txt to the wheel build to ensure consistent
building.

[1] http://logs.openstack.org/periodic/git.openstack.org/openstack/requirements/master/publish-wheel-mirror-centos-7/723e75b/python2/failed.txt

[2] https://github.com/requests/requests/issues/4673

Change-Id: I788994b69afd2769489454a3b16f84bea4c56e59