Skip released that are close but not fully EOLd when building wheels
The requirements project is very slow to actually tag branches as EOLd as
that can cause some testing environments and devstack to fail.  For
example if tox.ini refers to a constraints file in git[1].

However building wheels for those releases isn't actually helpful as
they're not chnaging *and* can cause failed runs as pypi etc move on.

This change adds an regex for the closed branches to exclude them from
building.  The regex is a little more complex than the single case
(newton) requires but allows for easier future additions.

[1] http://git.openstack.org/cgit/openstack/nova/tree/tox.ini?h=newton-eol#n12

Change-Id: Ia0c589989cb6e7e81e7b08e39d81dac78c55bf18