Remove Rally tagging ACL
The release process for official OpenStack components now uses
release requests using changes proposed to the openstack/releases
git repository. This allows peer review of release requests,
including version numbers, and run a series of checks to make
sure the release is likely to be successful once the tag is
pushed. The same process supports the creation of additional repo
branches as well. This all ensures that releases are reported
correctly on the releases.openstack.org website.

Rally is an official governed project and should be doing
all releases through the official release process. This
removes the ACL rights for the team to push their own tags.

Change-Id: I1f22fb731253059c20edd126fe0e7ac21666b40a