Bump packethost max servers to 100
This was originally set to 95 to ensure that we would have room to boot
test nodes if we needed them and to avoid conflict with our slightly
bigger than expected mirror host. However, we have quota for 100 and
nodepool has learned to dynamically check against quota so we should be
able to safely boot test instances and nodepool will avoid running up
against the quota accidentally.

As for the slightly larger than expected mirror node it is using more
disk which I don't think is one of the resource constraints we have to
worry about instead ram is our constraint but we are using only 8GB of
ram for the mirror node.

Change-Id: Ief5e8b923aab4f623c983c260c74d37fc3436836