Consolidate nodepool graphs
All of these dashboards are the same, and have mostly copied all the
same issues with them.  This makes updating anything a massive pain.

This implements a single dashboard template with a small script to
create individual dashboards for each provider and its regions.

I have included a range of fixes.  The y-axis format has changed in
later versions of grafana.  The API time tracking is no longer scaled,
but we just tell grafana it is in ms and it displays it correctly.
The test nodes history graph is moved to the top, as it is probably
the most interesting graph (note this splits itself out per region, if
mulitple regions are selected).  Values for "null as zero" are
consistently set.  Various formatting fixes for the labels are
included.

Change-Id: I5fbffaec3c82aa1fce0947f771de67edd15f7dfc