Use wildcard to apply system-required template
We require all projects to have the system-required template ... unless
we aren't gating them which is the case for ansible/ansible and
kata-containers/proxy. That's a bunch of boilerplate though, and zuul
now supports wildcard project configuration.

Make a wildcard config that applies to every project starting with
openstack that contains system-required. Remove the projects that didn't
have anything but that.

Change-Id: Ic1099625ae0500657e01e7cf48bce1273f7e0e24