Add fedora vexxhost flavor for Kata
The current plan for Kata is to show that rough parity with their
existing/previous CI works under Zuul and have them provide feedback on
that via third party check jobs run against kata-containers/runtime. If
this works out and Kata wants to move forward we will need to deploy
"production" Zuul CI for them in a new Zuul tenant.

Currently we only have Ubuntu Xenial testing for them, but they also
have Fedora and CentOS testing. Add a Fedora label to continue pushing
on this. Followup should be to add CentOS once Fedora is working, then
start commenting on the runtime repo.

Change-Id: Ie269c8a233ae4de8a3fe43c4f7d5ff11e1db1c82