Add Grafana dashboards for stable Neutron releases
In the Neutron CI meeting [1] it was discussed that failures on CI jobs
for stable branches were not getting sufficient visibility, because
the lack of dashboarding means that we are dealing with anecdotal data
as opposed to trended metrics.  This change adds "back one" and "back
two" dashboards (referring to stable/queens and stable/pike
respectively) that can persistently refer to the last two stable
releases, and will be updated with each release to refer to the proper
stable branches.

[1] http://eavesdrop.openstack.org/meetings/neutron_ci/2018/neutron_ci.2018-06-19-16.00.log.html#l-70

Change-Id: I8a372ffe640ef1125e6acc2d5e40ef54c9f8e261