increase timeout on neutron-lib periodic for vmware-nsx
At present the openstack-tox-py35-with-neutron-lib-master job for the
vmware-nsx project times out [1]. As a result we can't be sure if the
tests are fully successful or not and this impacts our perdicability
of neutron-lib releases.

This patch bumps the timeout. I'll talk to that team in regards to
slimming their UT suite.

[1] http://logs.openstack.org/periodic/git.openstack.org/openstack/
vmware-nsx/master/openstack-tox-py35-with-neutron-lib-master/0b29ffb/
job-output.txt.gz#_2018-06-27_07_08_05_006704

Change-Id: I43589f528c817006ffe7a7d393840ec27cb12ebd