Don't use requirements/constraints for infra docs
The infra docs publication jobs are sharing a base job with the
openstack ones. The problem is - Infra doesn't follow
requirements/constraints, so the job is actually building the docs for
publication inappropriately.

Split the base job in twain, allowing infra projects to appropriately
not be built with requirements/constraints applied.

Change-Id: I47050ca666a81a05ae85b956668c88848bb1a50b