Remove ironic-tempest-dsvm-ipa-wholedisk-bios-agent_ipmitool-tinyipa for nova
The job is overriding the irrelevant-files which means we end up
running it on nova docs-only changes.

We've moved the usage in-tree to the nova repo and fixed the
irrelevant-files so we don't need this defined in project-config
anymore. The in-tree change to nova will be backported to stable
branches to maintain coverage.

Depends-On: I9435fe84c8dc8a6f35ffb195ddbf71ca34edb242

Change-Id: I09e49512264909f1fa27f5d0e3703cdbb851ae4c
Related-Bug: #1776981