Revert "Add doc build job for Chef OpenStack"
Turns out that this is causing issues with stable branches that do not
yet have a doc/ tree. We'll define the docs builds in the local repo
instead, see [0].

Note that this isn't a clean revert, due to the removal of
"system-required" that has happened in the meantime, now the complete
project definition is removed.

[0] https://review.openstack.org/579148

This reverts commit 9c7980078e1101a432d925aea7bcdcc06e9fcffd.

Change-Id: I4e41e0a591e76483b036a3cc8e8c09a4d449a147