Make vexxhost specific pool for kata flavors
The kata test instances can only run in vexxhost currently and there are
delays in processing these node requests because we content with the
openstack requests as well. Split this into a separate pool so that we
can better model resource quotas/maxes as well as creating a separate
thread just for these node requests.

Change-Id: I449bba31c0e4a29d20af6ca7f3888bf9dc85945c