proposal/OSA: Sync the contents of the sync folder
We have an emerging need to synchronise a common set of tasks across
roles. While it may sometimes make sense to create a role for this
sort of thing, the overhead of doing so would be greater than simply
copying the file between roles. We're also finding include_role to
be somewhat buggy, especially when used from inside a role.

This patch adds the ability to copy a set of files from the source [1]
to the same path (without the 'sync' prefix) in the destination roles,
abiding by the same rules as all other synchronised files.

[1] openstack/openstack-ansible/tests/sync
Relates-to: https://review.openstack.org/579527
Change-Id: I598ff6aae50e6570d5f262c830c4be10f1cc0b75