Remove sahara-extra legacy jobs, moving in-tree
The legacy jobs for sahara-extra which builds the artifacts and
publishes them are going to be moved downstream, but better remove
them first from here so that the post jobs (legacy and new)
do not conflict with each other.
This means a short interval where sahara-extra will not have a build
job, but the amount of patches for sahara-extra is really small.

Needed-By: https://review.openstack.org/532691
Needed-By: https://review.openstack.org/581096
Change-Id: If924a8d2ee1de913452d81165993f94c31a42f6f