Reenable packethost in nodepool
Packethost instances have an mtu of 1500 now and we've tried to address
the small mtu problem in devstack gate and zuul jobs for multinode
network overlays. We should be good to start a small number of instances
in packethost to see that everything is working against before bumping
this number up to 100.

Change-Id: I52a73f039f4fe074042e367fa0262f0ce3715725