Run kata-runsh on the kata runtime repo
The runtime repo is far more active than the proxy repo. We used the
proxy repo job to get the job working. Now that the job runs
successfully [0] apply it to the runtime repo so that people will see
the results and we can start to receive feedback on the setup.

[0] https://github.com/kata-containers/proxy/pull/74#issuecomment-405294886

If kata is happy with this the next step will be deploying a kata zuul
tenant.

Depends-On: https://review.openstack.org/#/c/582289/
Change-Id: I16f508bd892112f7d24d43f8f9f3952552aa7229