Run propose-updates for requirements-constraints to Python3.6
Currently we're running propose-updates forrequirements-constraints
under python 3.5 (on xenial).  We take the results of freezing the
constraints fro 3.5 and emit them as 3.4 constraints also.  In the near
term[1] we'd like to use the same technique to generate 3.6 constraints.

With the inclusion of I8784509bc162eb6f2e80261bc2d81dbe63ce7989, and
knowing that we no longer run on trusty node, we can now clone the 3.5
constraints freeze to both 3.4 and 3.6.  However given that soon we'll
be moving from xenial to bionic this introduces debt that will need to
be fixed after the August release.

Instead just move propose-updates for requirements-constraints job to
bionic now and start cloning 3.6 to 3.5 and 3.4.

Several jobs use propose-updates as a parent so I've made this change
specifically where we call the propose-updates job specifically for
requirements-constraints

Note: This may feel very familiar if you reviewed
I541b64afaee3bf466ff1a461fd2cb472657c4986, The aim is the same but I
mis-understood what propose-update-constraints was doing and it wasn't
the right job and was reverted in I75280b23e2e60800af45dc2343c82008ec9971a9

[1] See re-design discussion:
    http://lists.openstack.org/pipermail/openstack-dev/2018-July/132152.html

Change-Id: I1874d4f20a6d69a881f0d9407e159c35ab21542d