Add trigger-readthedocs-webhook job
This job is to trigger readthedocs via their new API

The old API was quite simple in that anyone could hit the end-point
and trigger a new build.  The v2 API now requires authentication and a
unique id for each project we wish to update.

If projects wish, they can directly use the new "trigger-readthedocs"
role (from the dependent change) and create their own jobs, providing
their username/password or webhook authentication token directly from
a secret in their project playbook.

This job, however, uses the infra secret for the "openstackci" user to
update the docs.  Projects that wish to take this route require some
minor setup, but don't have to worry about managing new jobs or their
secrets and gain some benefits in having the openstackci user as a
backup admin.

To use the trigger-readthedocs-webhook job, projects will need to do
the following:

1) add openstackci user as an admin of their RTD project

2) generate an webhook via the "Integrations" dashboard page on the
   RTD project

3) provide the id from that webhook URL in job or project variables.
   This webhook id is not private; it is useless without
   authentication (it appears to be a global monotonically increasing
   integer for each webhook generated).

4) include the job. gerrit-dash-creator is populated with the new job
   for initial testing.  Other projects are currently using the
   "docs-on-readthedocs" template.  Change
   I3b65813671961d07c0a743685c537ad1df4bf68 proposes switching the
   template to use this new job.

Change-Id: Ic34ea5ead26a477a1cc5f8b25fd2ce1949c1dc13
Depends-On: https://review.openstack.org/579434
Depends-On: https://review.openstack.org/584230