Split roles after unbound into own play
This is because we actually want unbound to finish reloading before
running any other roles. Without this, we only restart unbound after
all roles have run.

This is to help debug an issue in fedora, where we cannot traceroute
to git.o.o properly, and think we are failing to resolve DNS properly.

  http://logs.openstack.org/47/586847/1/gate/windmill-fedora-27/2f06941/job-output.txt.gz

Change-Id: Ida150b3d484538caf0b040c417c391c7cd76dcc5
Signed-off-by: Paul Belanger <pabelanger@redhat.com>