Remove Ansible function jobs from cloud-launcher
The cloud-launcher needs a devstack deployment in order to be
able to run properly therefore the jobs listed never pass.

The jobs have been migrated and moved off to the repo but
those need to be removed first in order for the change that
adds the new working jobs to pass.

Change-Id: I45bd413bd468963be1cb11764291dad2f7bf2fbc
Needed-By: I5c714eb0732f345f56b9b0b9afa12ad1cc08fcdf