Bump packethost back to 100 max-servers
We've been running at 10 max servers for a day and a half and there are
no errors launching nodes according to grafana. Additionally the mirror
node (http://mirror.us-west-1.packethost.openstack.org) is still up and
running.

I think we are good to bump back up to 100 max servers given this under
the assumption that the disk size increases have made things happier.

Change-Id: I873c7067dc2f21e1f42b5b94fdec4e6df728e993