accessbot config: refactor
In responding to some comments on the original changes, it made me
realise that the global "alumni" and "mask" values are not actually
required.  Each channel looks up these values in the global dictionary
first, which is the correct place to put them.

This moves them into the global dictionary, but also adds and updates
(removes stackforge references) more details about what's happening
with the channel config keys.

Change-Id: Ic8fc5022d670e0dbcbb53ef803125317dfc008f7