New Airship project - Treasuremap
Airship is a collection of components that provide a declarative
interface for deploying bare metal Kubernetes at scale,
facilitating integrated deployment of OpenStack on Kubernetes.

The Airship Treasuremap project will serve these purposes:

1. House reference deployment manifests that can be used as a
starting point for setting up an Airship site.  The YAML
manifests will be self-documented so that the settings that
typically vary on a site-by-site basis can be easily found.

2. CICD pipelines that will integrate the reference manifests,
the various Airship components, and OpenStack (via OpenStack-Helm)
to ensure that the manifests are always up-to-date and that
changes to Airship and OpenStack-Helm do not result in an
integration regression.

3. A documentation overview of Airship, its components, and how
it can be used.

Change-Id: I0a8d98b544825e9be548031b0fd62d559f7d6722