Remove legacy-install-afs-with-puppet role
This role uses the system-config:install_puppet.sh script to first
install puppet on the wheel-builder host, and then uses
puppet-kerberos/puppet-openafs to bring up the AFS environment.

This is replaced with ansible roles to do the same AFS bring-up.  A
new "openafs-client" playbook is added to configure this for the infra
environment.  The wheel-building base job is reworked to use this.

Roles have been tested under zuul with
https://review.openstack.org/596018

kerberos-client role:
Depends-On: https://review.openstack.org/593486
openafs-client role:
Depends-On: https://review.openstack.org/593487
add system-config to linters:
Depends-On: https://review.openstack.org/596645
Change-Id: I9a3af36048e4a4439fe86b36582667ff0af68035