Bump ansible for linters
I noticed this when I tried to use Ansible 2.5 "loop:" constructs.

Unfortunately, we can't just rely on bringing in zuul to pull the
right version of ansible.  ansible-lint being uncapped just takes it
over, and we end up with the latest version.  Pin it to the current
zuul requirements and add a note (maybe one day
https://github.com/pypa/pip/issues/988 will get a fix ...)

Change-Id: Iaf8f0f5cdc46df41fec7c436c1179c80bb5c368e