drop python 3.5 jobs from oslo.service
The python 3.5 jobs are failing on older stable branches where we
never successfully supported that version. We have several other
changes lined up to import the test settings into the oslo.service
repo, but we can't land those while the 3.5 tests are still
running. This patch removes the template that causes the jobs to run.

We will restore the tests on the branches where they work in those
separate patches, and come back with a later patch to remove all of
the rest of the job settings (we don't want the repo completely
untested during the migration).

Change-Id: I897e198898dd5d005b00038a65a0a214428ee63f
Signed-off-by: Doug Hellmann <doug@doughellmann.com>