Add graphs with overall number of jobs run
For each graph in Neutron failure rate dashboard in Grafana
there is now additional graph which shows total number of jobs
run in last 24h.
With that information we will be able to know better if
e.g. spike of failure rate of some job type is really problem
in this job or maybe it's just for example problem of very small
number of runs of such job.
Let's say that UT is suddenly spike to 100% of failures - if that
is because there were only 2 runs of UT recently and those 2 times
fails, it's probably not big problem from CI point of view.
But if this spike is seen on e.g. 10 runs - it probably means that
something is wrong with this job and we should definitely check that.

Change-Id: I5cf81301eb8a4f16a18dedeb317d12d88def9465