drop new job templates for oslo.messaging
The python 3.5 and documentation PTI jobs do not work in the older
branches of oslo.messaging, which makes moving the zuul configuration
in-tree impossible. This patch drops the configuration for those jobs
here so we can land the other patches to add the relevant jobs in
those branches. The follow-up patch will finish removing the rest of
the settings.

Change-Id: I9be334ed175812bc0cf24d9fddbf8fd017d70309
Signed-off-by: Doug Hellmann <doug@doughellmann.com>