Neutron failure rate - include TIMED_OUT jobs on graphs
As TIMED_OUT jobs are usually related to issues in slow test/api
responses it is worth to include them in graphs to see how many jobs
are failing because of timeouts.

Since now TIMED_OUT jobs will be counted in same way as FAILURES
in Neutron failure rate Grafana dashboard.

Change-Id: Id51f3f3ecae3e4a41281a46861e1ffc5f549b72c