Add virtualbmc-core group
In order to grant repository level permissions to a new
core reviewer for virtualbmc, we realized that we need
a group for that purpose like our other repositories.

As such, requesting the addition of virtualbmc-core group
such that we can manage permissions for the repository and
grant core reviewer powers without requiring changing the
ironic-core group.

After merging, we will naturally need ironic-core added to
the group, so we can orchustrate/change the permissions with
in the group and add the new reviewer as a group member.

Change-Id: I16eab9b5fddabc2c8292a6d740e5ddc595275795