Match older neutron dashboards to HEAD one, bump versions
Neutron has test failure dashboards not just for the current HEAD, but
for the most recent release and the one before that.  The HEAD dashboard
was revamped, and this brings the dashboards for the stable releases
into line with that.  Also it bumps the versions now that stable/rocky
is about to be the current release.

Change-Id: Ic06660e32a37643420b610529f0eb34938c9887c