wheel-build: only install afs client for publish jobs
The base build jobs don't need to interact with AFS, so remove the
client install.

The publish jobs run the wheel builds then but copy the files to the
AFS volumes.  Add the openafs-client role to these.

The release jobs run after the publish jobs and essentially do a "vos
release" on the volumes that were just populated by the publish jobs.
These run on the executor (which is already an AFS client) so nothing
to do here.

Change-Id: Iaa7c2ad23930fcdbd10a952d11e1647fa6e57b3b