wheel-build: ignore existing files and fix error log
It seems that on trusty, when two branches have the same version of a
requirement the first version builds, but then pip explodes on the
second try as it prompts you asking if you want to overwrite the
existing download.  Use the "--exists-action" flag to ignore.

We have added two arguments for the upper-constraints.txt, which has
pushed out the columns for the error matching.  Fix that too.

Change-Id: I53ed7d8a86e6910b00a181cc3ca37b7aa4b7bb48