wheel-index: only create index once
We saw what looks very much like the index generation racing, with
"mv" unable to close the index file ...

Move to a linear strategy so that each host finishes each task before
moving on.  Then add run_once to the index generation so only one of
the hosts in a python2/3 situation creates the index file.

Change-Id: I3761dd9d2ab752e66bcccdd2434405f11535b95f