switch release job for networking-bgpvpn
The old release job uses tox to run setuptools commands. This requires
that the dependencies for the package being released also be
installable with pip.

The networking-bgpvpn package depends on networking-odl which depends
on a git URL for ceilometer, which is not on PyPI and therefore cannot
be installed with pip.

In order to unblock the rocky release for networking-bgpvpn, we go
ahead and switch the release job template to the new form of the job,
which uses the newer PTI and does not require tox to build the
package.

Change-Id: I6028fc9b83b6d40df4dcb552284c8f2d68d2fdb8
Signed-off-by: Doug Hellmann <doug@doughellmann.com>