Create an openstack/placement project
This patch includes the changes necessary to bootstrap an
openstack/placement repo that contains code seeded
from https://github.com/EdLeafe/placement . That repo contains
a clone of openstack/nova filtered by the filter_git_history.sh
script from oslo.tools [1]. The result is a collection of the
files that make up the placement service, it's test and docs
and auxilliary stuff. It is not a working repo. It will be
made to work via a series of changes to be reviewed in gerrit
once the new repo is seeded. Initially those changes will be
fairly mechanical, to fix file and import paths, and to remove
unneeded files that were included in the results because they
were non-placement members of commit required in the filtering.

The first commit will add a .zuul.yaml file that sets up some
noop-jobs and non-voting versions of openstack-tox-functional,
openstack-tox-functional-py35, openstack-tox-py27, openstack-tox-py35,
openstack-tox-p27, openstack-tox-pep8. As changes to the gerrit repo
progress, the jobs will be made voting and additional jobs will be added
for docs, api-ref, etc. Eventually the configuration will be update
to use templates.

The placement-core gerrit group will initially be seeded with the
existing nova-core group and a small number of additional people
selected from existing placement contributors.

The openstack/placement repo will initially be a part of the official
nova project. A patch indicating those plans will be made to governance.
It will depend on this change.

[1] https://git.openstack.org/cgit/openstack/oslo.tools/tree/filter_git_history.sh

Needed-By: https://review.openstack.org/598380
Change-Id: Icab8e931e5ae589eab0038ca12e6b5bdba04ff43