Change neutron grafana dashboard to 2-column format
The new layout for the Neutron grafana dashboard expresses results in
sets of two: first the rate of failure, then the number of jobs.  But
the display we had been using was formatted for 3 columns, which meant
things did not line up.  I experimented using the "older release"
neutron dashboards with increasing the span from 4 to 6, which resizes
from three columns to 2.  This means that all of the failure rate graphs
are in one column, and all the job count graphs are in a second column,
for easy comparison.  This change implements the 2-column format for the
main neutron dashboard.

Here are links so you can compare the format:

Main dashboard: http://grafana.openstack.org/d/Hj5IHcSmz/neutron-failure-rate?orgId=1
Older dashboard: http://grafana.openstack.org/d/pM54U-Kiz/neutron-failure-rate-previous-stable-release?orgId=1

Change-Id: I51d7f9ae895cf7ef4caace5a74cf66f78e3581cf