Stop requiring users be registered on openstack channels
Freenode sustained a spam attack that forced us to require users be
registered on our channels. Recently we removed this requirement on
the #openstack-infra channel to see if the spam attacks have persisted.
Now a day later we have had no spam attacks.

Remove the +r (and related +f) modes from our channels to be more
friendly to users.

Change-Id: I10783a393765f14e6dc44be46b789b67fdbbc7f3