Move coverage jobs to voting check job
There was a thread some time ago that suggested coverage jobs weren't
doing much in the "post" pipeline because nobody looks at them and the
change numbers may be difficult to find anyway [1].  This came up
again in a cleanup to add non-voting coverage jobs in
I5c42530d1dda41b8dc8c13cdb10458745bec7bcc

There really is no consistency across projects; it seems like a couple
of different approaches have been cargo-cult copied as new projects
came in, depending on which random project was used as a template.

This change does a cleanup by moving all post coverage jobs into the
check queue as voting jobs using the template "openstack-cover-jobs".
For infra repos that use "tox-cover", the set up is changed to use a
voting check job.

Note that only repos are updated that are not currently in the process
of the python3-first migration, these repos need separate handling
in-repo.

[1] http://lists.openstack.org/pipermail/openstack-dev/2016-July/099491.html

Co-Authored-By: Andreas Jaeger <jaegerandi@gmail.com>
Co-Authored-By: Ian Wienand <iwienand@redhat.com>
Change-Id: I4e64bb803929f1dbaef26bf513d3a0afa91484d8