Revert "add signing steps to javascript publishing job"
Turns out playbooks/javascript/publish.yaml is also used in
generating branch tarballs, which we don't want to sign. This has
resulted in the publish-openstack-javascript-content job, e.g. for
the openstack-infra/storyboard-webclient project which runs it after
every commit, to fail with:

The task includes an option with an undefined variable. The error
was: 'gpg_key' is undefined

We probably need to switch to running separate versions of this job
for tags and branch updates so that we can exclusively sign the
former.

This reverts commit 7244f69975311ac8f570082e419d45563ca7cc16.

Change-Id: I77fd1bd315d96f8fe082b3e50a430cd8602c4d71