remove job settings for Quality Assurance repositories
This is a mechanically generated patch to remove the job settings that
have been migrated to the git repositories owned by the Quality Assurance team.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

This updates legacy-periodic-tempest-dsvm-oslo-latest-full-master as
well to mark it running on master and removes unused irrelevant-files
for a periodic job.

Change-Id: I6c63fef537800165cb44c1a644f90b565edf9115
Story: #2002586
Task: #24327