Disable unbound chroot on test node images
On centos this seems to be part of the built in config, on ubuntu this
seems to not be configured but is unbounds default behavior there, on
gentoo it chroots per the default described in the manpage.

In order to force consistent behavior across all systems disable the
chroot when we configure our logfile path (so that the logfile's dir is
present and writeable) by setting chroot: "" across the board.

This should fix the unbound service on gentoo and be a noop for our
other platforms.

Change-Id: Ic1b66c1982c14759e8fd8370452df21d2b0d9510