Add new project: sardonic
Sardonic is a hardware inventory system; not quite a CMDB, but a high
quality data source for input into a CMDB.

The idea for sardonic came out of some conversations between Oath and
CERN, and we wrote up a document describing some of the use cases,
requirements, and other random thoughts:
https://docs.google.com/document/d/144g8E_fzGD4WZzMvswkeowzcILL4hxg4QDS-46KYCcQ/edit
(apologies for the google doc - this will likely become a document
within sardonic's documentation)

This was presented to the ironic team during the Stein PTG, and folks
were generally receptive to the idea. Some people from Red Hat and Dell
were also interested in contributing, and we decided the next steps
would be to spin up a repository in the OpenStack infrastructure.

The long-term goal is for sardonic to integrate with ironic, and
potentially other projects that deal with hardware (directly or
indirectly), like Nova or Neutron.

It's unclear if sardonic should be managed within the Ironic project, as
a separate official OpenStack project, or remain unofficial; so we treat
it as if it will eventually be official to reduce friction as much as
possible.

Change-Id: Ia7f339d1effbf650d6adfe9ddb94fac52d38a0fd