Revert "Disable OVH BHS1 region"
This reverts commit 19e7cf09d9a9aa4730faca3f3cc506fed0cb454e.

The issues in OVH BHS1 around networking configuration have been worked
around with updates to glean and configuration to the labels in zuul.
New images are in place for each supported image in BHS1. We can go
ahead and start using this region again.

I have manually tested this by booting an ubuntu-xenial node with
glean_ignore_interfaces='True' set in metadata and the networking comes
up with expected using DHCP. The mirror in that region is reachable from
this test node.

Change-Id: I29746686217a62709c4afc6656d95829ace6fb3b