Grafana: set zuul node requests yaxis min
Since node requests are something that we generally want to see
reduced to 0, it can be misleading to show a graph where the lowest
point on the y axis is, for example (as I write this) nearly 4,000.
Fix the minimum to zero so it's easier to see what the overall trend
toward zero is.

Change-Id: Iad1b5667fd6d1d4bfa9fe50706ad71debd01d5c6