Align neutron grafana with actual jobs being run
There has been some drift between the neutron grafana dashboard
configuration and what jobs are actually being run.  This fixes a number
of issues:

1.) Adds the following jobs that were missing:
        - openstack-tox-pep8
        - openstack-tox-py36
        - tempest-full-py3
        - build-openstack-releasenotes (Check queue only)
        - openstack-tox-docs

2.) Renames the job "neutron-tempest-multinode-full" to
"tempest-multinode-full" to reflect how it is now named.

3.) Remove these jobs that no longer exist or are being removed:
        - legacy-tempest-dsvm-py35
        - neutron-fullstack-python35

Also, in Grafana there is no line for the tempest-slow job, despite the
fact that it is defined here and it is clearly running in the gate.  So
I change it's position slightly to try to get it to populate.

Change-Id: I2c47742630f959024dcf8e647373ccc913ceb6de